requirement.txt
# requirements.txt
streamlit>=1.28.0
agno>=0.1.0
icalendar>=5.0.0
asyncio>=3.4.3
requests>=2.31.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2